eatgo-common/src/main/java/kr/co/fastcampus/eatgo/utils/JwtUtil.java
package kr.co.fastcampus.eatgo.utils;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;
import io.jsonwebtoken.security.Keys;
import org.springframework.stereotype.Component;

import java.security.Key;

//@Component
public class JwtUtil {

    private Key key;

    public JwtUtil(String secret) {
        this.key = Keys.hmacShaKeyFor(secret.getBytes());
    }

    public String createToken(long userId, String name) {
//        String secret = "12345678901234567890123456789012";
//        Key key = Keys.hmacShaKeyFor(secret.getBytes());

        return Jwts.builder()
                .claim("userId", userId)
                .claim("name", name)
                .signWith(key, SignatureAlgorithm.HS256)
                .compact();
    }

    public Claims getClaims(String token) {
        // 위에서도 sign key 를 잡아준것처럼 아래도 잡아준다.
        return Jwts.parser()
                .setSigningKey(key)
                .parseClaimsJws(token) // jws: 사인이 포함된 jwt 를 의미
                .getBody();
    }
}
